activities/requirements.txt
black==23.9.1
click==8.1.7
iniconfig==2.0.0
isort==5.12.0
mypy==1.5.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
pluggy==1.3.0
protobuf==4.24.4
pytest==7.4.2
pytest-asyncio==0.21.1
python-dateutil==2.8.2
six==1.16.0
temporalio==1.3.0
types-protobuf==4.24.0.2
typing_extensions==4.8.0
requests==2.31.0